pyproject.toml
[project]
name = "django-job"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.4",
    "django-allauth>=65.10.0",
    "django-crispy-forms>=2.4",
    "django-debug-toolbar>=6.0.0",
    "django-unfold>=0.63.0",
]
